+++ app/CouponReturn.php
@@ -4,9 +4,19 @@ namespace App;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Auth;
+use OwenIt\Auditing\Contracts\Auditable;
 
-class CouponReturn extends Model
+class CouponReturn extends Model implements Auditable
 {
+    use \OwenIt\Auditing\Auditable;
+
+    protected $fillable = [
+        'project_id',
+        'coupon_type_id',
+        'date',
+        'amount',
+        'user_id',
+    ];
 
     public static function boot()
     {
@@ -16,14 +26,6 @@ class CouponReturn extends Model
 
         parent::boot();
     }
-
-    protected $fillable = [
-        'project_id',
-        'coupon_type_id',
-        'date',
-        'amount',
-        'user_id',
-    ];
 
     public function couponType() {
         return $this->belongsTo('App\CouponType');
@@ -35,5 +37,16 @@ class CouponReturn extends Model
 
     public function user() {
         return $this->belongsTo('App\User');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function generateTags(): array
+    {
+        return [
+            $this->project->name,
+            $this->couponType->name,
+        ];
     }
 }